Updated OAISSN4
Added ability to process Journal's with over 1000 DOIs in a year. 

Added option to export details about the best OA link for each DOI to a csv file.

Also added a line of code that will check if journal is Gold OA, if so, it will let the user know.